Added js and php validation for upload and edit video

+++ controller/VideoController.php
@@ -19,22 +19,22 @@ class VideoController extends BaseController {
     public function uploadAction() {
         try {
             $videoDao = VideoDao::getInstance();
+            $tagDao = TagDao::getInstance();
             $requestMethod = $_SERVER['REQUEST_METHOD'];
+            $tags = $tagDao->getAll();
+            $thumbnailUrl = null;
+            $title = null;
+            $description = null;
+            $videoUrl = null;
+            $pageTitle = 'Upload video';
+            $btnText = 'Upload';
+            $previewDivDisplay = 'none';
+            $fileInputFieldDisplay = 'block';
+            $required = 'required';
+            $formAction = 'index.php?page=upload';
+            $selected = 16;
 
             if ($requestMethod == 'GET') {
-                $tagDao = TagDao::getInstance();
-                $tags = $tagDao->getAll();
-                $thumbnailUrl = null;
-                $title = null;
-                $description = null;
-                $videoUrl = null;
-                $pageTitle = 'Upload video';
-                $btnText = 'Upload';
-                $previewDivDisplay = 'none';
-                $fileInputFieldDisplay = 'block';
-                $required = 'required';
-                $formAction = 'index.php?page=upload';
-                $selected = 'Other';
                 $this->render('video/upload', [
                     'tags' => $tags,
                     'thumbnail_url' => $thumbnailUrl,
@@ -52,20 +52,44 @@ class VideoController extends BaseController {
             } elseif ($requestMethod == 'POST') {
                 if (isset($_SESSION['user']) &&
                     isset($_POST['title']) &&
-                    isset($_POST['description'])
-                ) {
+                    isset($_POST['description'])) {
 
-                    $resultMsg = 'Your video was successfully uploaded!';
-                    $userId = $_SESSION['user']->getId();
-
+                    $title = $_POST['title'];
+                    $description = $_POST['description'];
                     $tmpFileName = $_FILES['video-file']['tmp_name'];
                     $realFileName = $_FILES['video-file']['name'];
+                    $tagId = $_POST['tags'];
+                    $userId = $_SESSION['user']->getId();
+
+                    $resultMsg = 'Your video was successfully uploaded!';
                     $fileType = $_FILES['video-file']['type'];
                     $videoId = '';
-                    if (is_uploaded_file($tmpFileName)) {
-                        if (strpos($fileType, 'video') === false) {
-                            $resultMsg = 'Error! File is not a video!';
-                        } else {
+
+                    $validator = Validator::getInstance();
+                    $validTitle = $validator->validateTitle($title);
+                    $validDescription = $validator->validateDescription($description);
+                    $extensions = ['mp4', 'webm', 'ogg'];
+                    $validVideo = $validator->validateUploadedVideo($realFileName, $tmpFileName, 52428800, 'video', $extensions);
+
+                    $errors = array();
+                    if (is_array($validTitle)) {
+                        foreach ($validTitle as $error) {
+                            $errors['title'][] = $error;
+                        }
+                    }
+                    if (is_array($validDescription)) {
+                        foreach ($validDescription as $error) {
+                            $errors['description'][] = $error;
+                        }
+                    }
+                    if (is_array($validVideo)) {
+                        foreach ($validVideo as $error) {
+                            $errors['video'][] = $error;
+                        }
+                    }
+
+                    if (empty($errors)) {
+                        if (is_uploaded_file($tmpFileName)) {
                             if (!file_exists('../uploads/videos')) {
                                 mkdir('../uploads/videos', 0777);
                             }
@@ -73,45 +97,68 @@ class VideoController extends BaseController {
                             $videoPath = "../uploads/videos/$videoName." . pathinfo($realFileName, PATHINFO_EXTENSION);
                             $thumbPath = "../uploads/thumbnails/$videoName.png";
                             move_uploaded_file($tmpFileName, $videoPath);
-                            if (file_exists($videoPath)) {
-
-                                if (!file_exists('../uploads/thumbnails')) {
-                                    mkdir('../uploads/thumbnails', 0777);
-                                }
-                                file_put_contents($thumbPath, file_get_contents('data://' . $_POST['thumbnail']));
-                                $newVideo = new Video(
-                                    null,
-                                    $_POST['title'],
-                                    $_POST['description'],
-                                    date('Y-m-d'),
-                                    $userId,
-                                    $videoPath,
-                                    $thumbPath,
-                                    $_POST['tags'],
-                                    '0'
-                                );
-                                try {
-                                    $videoId = $videoDao->insert($newVideo);
-                                } catch (\PDOException $e) {
-                                    if (file_exists($videoPath)) {
-                                        unlink($videoPath);
-                                    }
-                                    if (file_exists($thumbPath)) {
-                                        unlink($thumbPath);
-                                    }
-                                    $resultMsg = 'An error occurred! Please try again later.';
-                                }
+                            if (!file_exists('../uploads/thumbnails')) {
+                                mkdir('../uploads/thumbnails', 0777);
                             }
+                            file_put_contents($thumbPath, file_get_contents('data://' . $_POST['thumbnail']));
+                            $newVideo = new Video(
+                                null,
+                                $_POST['title'],
+                                $_POST['description'],
+                                date('Y-m-d'),
+                                $userId,
+                                $videoPath,
+                                $thumbPath,
+                                $tagId,
+                                '0'
+                            );
+                            try {
+                                $videoId = $videoDao->insert($newVideo);
+                                $this->render('video/upload-result', [
+                                    'result' => $resultMsg,
+                                    'video_id' => $videoId,
+                                    'success' => 1
+                                ]);
+                            } catch (\PDOException $e) {
+                                if (file_exists($videoPath)) {
+                                    unlink($videoPath);
+                                }
+                                if (file_exists($thumbPath)) {
+                                    unlink($thumbPath);
+                                }
+                                $resultMsg = 'An error occurred! Please try again later.';
+                                $this->render('video/upload-result', [
+                                    'result' => $resultMsg,
+                                    'video_id' => $videoId,
+                                    'success' => 0
+                                ]);
+                            }
+                        } else {
+                            $resultMsg = 'Error uploading file!';
+                            $this->render('video/upload-result', [
+                                'result' => $resultMsg,
+                                'video_id' => $videoId,
+                                'success' => 0
+                            ]);
                         }
                     } else {
-                        $resultMsg = 'Error uploading file!';
+                        $this->render('video/upload', [
+                            'tags' => $tags,
+                            'thumbnail_url' => $thumbnailUrl,
+                            'title' => $title,
+                            'description' => $description,
+                            'video_url' => $videoUrl,
+                            'page_title' => $pageTitle,
+                            'btn_text' => $btnText,
+                            'preview_div_display' => $previewDivDisplay,
+                            'file_input_display' => $fileInputFieldDisplay,
+                            'required' => $required,
+                            'form_action' => $formAction,
+                            'selected' => $tagId,
+                            'errors' => $errors
+                        ]);
                     }
-                    $this->render('video/upload-result', [
-                        'result' => $resultMsg,
-                        'video_id' => $videoId
-                    ]);
-                } else {
-                    header('Location: index.php?page=upload');
+
                 }
             }
         }
@@ -308,23 +355,23 @@ class VideoController extends BaseController {
         try {
             $requestMethod = $_SERVER['REQUEST_METHOD'];
             $videoDao = VideoDao::getInstance();
-            if ($requestMethod == 'GET') {
-                $tagDao = TagDao::getInstance();
-                $tags = $tagDao->getAll();
-                $videoId = $_GET['id'];
-                $video = $videoDao->getByID($videoId);
-                $thumbnailUrl = $video->getThumbnailURL();
-                $title = $video->getTitle();
-                $description = $video->getDescription();
-                $videoUrl = $video->getVideoURL();
-                $tagId = $video->getTagId();
-                $pageTitle = 'Edit video';
-                $btnText = 'Edit';
-                $previewDivDisplay = 'block';
-                $fileInputFieldDisplay = 'none';
-                $required = '';
-                $formAction = 'index.php?page=edit-video';
+            $tagDao = TagDao::getInstance();
+            $tags = $tagDao->getAll();
+            $videoId = isset($_GET['id']) ? $_GET['id'] : $_POST['video-id'];
+            $video = $videoDao->getByID($videoId);
+            $thumbnailUrl = $video->getThumbnailURL();
+            $title = $video->getTitle();
+            $description = $video->getDescription();
+            $videoUrl = $video->getVideoURL();
+            $tagId = $video->getTagId();
+            $pageTitle = 'Edit video';
+            $btnText = 'Edit';
+            $previewDivDisplay = 'block';
+            $fileInputFieldDisplay = 'none';
+            $required = '';
+            $formAction = 'index.php?page=edit-video';
 
+            if ($requestMethod == 'GET') {
                 $this->render('video/upload', [
                     'video_id' => $videoId,
                     'tags' => $tags,
@@ -345,27 +392,65 @@ class VideoController extends BaseController {
                 isset($_POST['video-id']) &&
                 isset($_POST['title']) &&
                 isset($_POST['description']) &&
-                isset($_POST['old-thumbnail-url'])
-            ) {
+                isset($_POST['old-thumbnail-url'])) {
 
-                $thumbnailURL = $_POST['old-thumbnail-url'];
-                if (file_exists($thumbnailURL) && $_POST['thumbnail'] != '') {
-                    unlink($thumbnailURL);
-                    file_put_contents($thumbnailURL, file_get_contents('data://' . $_POST['thumbnail']));
+                $validator = Validator::getInstance();
+                $title = $_POST['title'];
+                $description = $_POST['description'];
+                $validTitle = $validator->validateTitle($title);
+                $validDescription = $validator->validateDescription($description);
+
+                $errors = array();
+                if (is_array($validTitle)) {
+                    foreach ($validTitle as $error) {
+                        $errors['title'][] = $error;
+                    }
                 }
-                $editedVideo = new Video($_POST['video-id'], $_POST['title'], $_POST['description'], null, null, null, null, $_POST['tags'], '0');
-                try {
-                    $videoDao = VideoDao::getInstance();
-                    $rowsAffected = $videoDao->edit($editedVideo);
-                    $resultMsg = 'Video edited successfully!';
-                    if ($rowsAffected == 1) {
+                if (is_array($validDescription)) {
+                    foreach ($validDescription as $error) {
+                        $errors['description'][] = $error;
+                    }
+                }
+
+                if (empty($errors)) {
+                    $thumbnailURL = $_POST['old-thumbnail-url'];
+                    if (file_exists($thumbnailURL) && $_POST['thumbnail'] != '') {
+                        unlink($thumbnailURL);
+                        file_put_contents($thumbnailURL, file_get_contents('data://' . $_POST['thumbnail']));
+                    }
+                    $editedVideo = new Video($_POST['video-id'], $_POST['title'], $_POST['description'], null, null, null, null, $_POST['tags'], '0');
+                    try {
+                        $videoDao->edit($editedVideo);
+                        $resultMsg = 'Video edited successfully!';
                         $this->render('video/upload-result', [
                             'result' => $resultMsg,
                             'video_id' => $_POST['video-id']
                         ]);
+                    } catch (\PDOException $e) {
+                        $resultMsg = 'An error occurred! Please try again later!';
+                        $this->render('video/upload-result', [
+                            'result' => $resultMsg,
+                            'video_id' => $_POST['video-id'],
+                            'success' => 0
+                        ]);
                     }
-                } catch (\PDOException $e) {
-                    $resultMsg = 'An error occurred! Please try again later!';
+                } else {
+                    $this->render('video/upload', [
+                        'video_id' => $videoId,
+                        'tags' => $tags,
+                        'thumbnail_url' => $thumbnailUrl,
+                        'title' => $_POST['title'],
+                        'description' => $_POST['description'],
+                        'video_url' => $videoUrl,
+                        'page_title' => $pageTitle,
+                        'btn_text' => $btnText,
+                        'preview_div_display' => $previewDivDisplay,
+                        'file_input_display' => $fileInputFieldDisplay,
+                        'required' => $required,
+                        'form_action' => $formAction,
+                        'tag' => $tagId,
+                        'errors' => $errors
+                    ]);
                 }
 
             }
